Update exercise 19.cpp

+++ Computer Science/CM2005 Object-Oriented Programming/Book Code/Programming Exercises/Chapter 2/exercise 19.cpp
@@ -5,10 +5,10 @@ using namespace std;
 int main()
 {
 
-  double quarters;
-  double dimes;
-  double nickels;
-  double pennies;
+  int quarters;
+  int dimes;
+  int nickels;
+  int pennies;
 
   cout << "Enter quarters: " << endl;
   cin >> quarters;
